Cooperative-scheduler: Print a dot during a timer interrupt
