+++ src/interrupts/mod.rs
@@ -1,6 +1,6 @@
 use crate::scheduler::task::keyboard::{scancode_stream::SCANCODE_QUEUE, KEYBOARD_WAKER};
 use crate::utils::Port;
-use crate::{print, println};
+use crate::{print, println, timer_print};
 use lazy_static::lazy_static;
 use x86_64::structures::idt::{InterruptDescriptorTable, InterruptStackFrame, PageFaultErrorCode};
 
@@ -73,7 +73,7 @@ extern "x86-interrupt" fn page_fault_handler(
 }
 
 extern "x86-interrupt" fn timer_interrupt_handler(_frame: InterruptStackFrame) {
-    print!(".");
+    timer_print!(".");
     PICS.lock().end_of_interrupt(Interrupts::Timer as u8);
 }
